Rename to CODEM, bump version
