--- arcgis/3D Data Co-Registration.Register_MultiType.pyt.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0"?>
-<metadata xml:lang="en"><Esri><CreaDate>20210815</CreaDate><CreaTime>11573000</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce><ModDate>20220613</ModDate><ModTime>15095700</ModTime><scaleRange><minScale>150000000</minScale><maxScale>5000</maxScale></scaleRange><ArcGISProfile>ItemDescription</ArcGISProfile></Esri><tool name="Register_MultiType" displayname="Register Multi-Type" toolboxalias="3d_registration" xmlns=""><arcToolboxHelpPath>c:\program files\arcgis\pro\Resources\Help\gp</arcToolboxHelpPath><parameters><param name="foundation_file" displayname="Foundation Data File" type="Required" direction="Input" datatype="File" expression="foundation_file"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Path to foundation data file.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="aoi_file" displayname="Area of Interest (AOI) Data File" type="Required" direction="Input" datatype="File" expression="aoi_file"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Path to area of interest data file.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="min_resolution" displayname="Minimum Resolution (m)" type="Required" direction="Input" datatype="Double" expression="min_resolution"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Minimum allowable registration pipeline data resolution. Smaller values increase computation time.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_solve_scale" displayname="DSM Registration" type="Required" direction="Input" datatype="Boolean" expression="dsm_solve_scale"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Flag to include or exclude scale from the solved coarse registration transformation.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_solve_scale" displayname="ICP Registration" type="Required" direction="Input" datatype="Boolean" expression="icp_solve_scale"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Flag to include or exclude scale from the solved fine registration transformation.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_strong_filter" displayname="Normalization Strong Filter Size (m)" type="Required" direction="Input" datatype="Double" expression="dsm_strong_filter"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Standard deviation of the large Gaussian filter used to normalize the DSM prior to feature extraction. Larger values allow longer wavelength vertical features to pass through into the normalized DSM.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_weak_filter" displayname="Normalization Weak Filter Size (m)" type="Required" direction="Input" datatype="Double" expression="dsm_weak_filter"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Standard deviation of the small Gaussian filter used to normalize the DSM prior to feature extraction. Larger values increasingly blur short wavelength vertical features in the normalized DSM.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_akaze_threshold" displayname="AKAZE Detection Threshold" type="Required" direction="Input" datatype="Double" expression="dsm_akaze_threshold"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;AKAZE feature detection response threshold. Larger values require increasingly distinctive local geometry for a feature to be detected.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_lowes_ratio" displayname="Lowe's Ratio" type="Required" direction="Input" datatype="Double" expression="dsm_lowes_ratio"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Feature matching relative strength control. Larger values allow weaker matches relative to the next best match.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_ransac_max_iter" displayname="RANSAC Maximum Iterations" type="Required" direction="Input" datatype="Long" expression="dsm_ransac_max_iter"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Maximum number of iterations allowed in the RANSAC algorithm that filters the putative feature matches.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="dsm_ransac_threshold" displayname="RANSAC Error Threshold (m)" type="Required" direction="Input" datatype="Double" expression="dsm_ransac_threshold"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;P&gt;&lt;SPAN&gt;Maximum residual error for a matched feature pair to be included in a RANSAC solution to a 3D registration transformation. Larger values include matched feature pairs with increasingly greater disagreement with the solution.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_max_iter" displayname="ICP Maximum Iterations" type="Required" direction="Input" datatype="Long" expression="icp_max_iter"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Maximum number of iterations allowed in the iterative closest point (ICP) algorithm&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_angle_threshold" displayname="ICP Angle Threshold (degrees)" type="Required" direction="Input" datatype="Double" expression="icp_angle_threshold"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;ICP convergence criterion: minimum change in Euler angle between iterations.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_distance_threshold" displayname="ICP Distance Threshold (m)" type="Required" direction="Input" datatype="Double" expression="icp_distance_threshold"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;P&gt;&lt;SPAN&gt;ICP convergence criterion: minimum change in translation between iterations.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_rmse_threshold" displayname="ICP RMSE Relative Change Threshold" type="Required" direction="Input" datatype="Double" expression="icp_rmse_threshold"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;P&gt;&lt;SPAN&gt;ICP convergence criterion: minimum relative change between iterations in the root mean squared error.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;</dialogReference></param><param name="icp_robust" displayname="Robust ICP" type="Required" direction="Input" datatype="Boolean" expression="icp_robust"><dialogReference>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Flag to include or exclude robust weighting in the fine registration solution.&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</dialogReference></param></parameters><summary>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;Toolbox to perform multi-type registration&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</summary><usage>&lt;DIV STYLE="text-align:Left;"&gt;&lt;DIV&gt;&lt;P&gt;&lt;SPAN&gt;test&lt;/SPAN&gt;&lt;/P&gt;&lt;/DIV&gt;&lt;/DIV&gt;</usage></tool><dataIdInfo><idCitation><resTitle>Register Multi-Type</resTitle></idCitation><searchKeys><keyword>registration</keyword></searchKeys></dataIdInfo><distInfo><distributor><distorFormat><formatName>ArcToolbox Tool</formatName></distorFormat></distributor></distInfo><mdHrLv><ScopeCd value="005"/></mdHrLv><mdDateSt Sync="TRUE">20220613</mdDateSt></metadata>
