+++ arcgispro/src/setup.py
@@ -3,8 +3,8 @@ from setuptools import setup
 
 setup(
     name="coregister",
-    version="0.21",
-    author="Preston Hartzell",
+    version="0.22",
+    author="Ognyan Moore",
     description="An ArcGIS Pro toolbox for 3D data co-registration",
     packages=["coregister"],
     package_data={
