+++ readme.md
@@ -1,60 +1,57 @@
-# MODEM: Multi-Modal Digital Elevation Model Registration
+# CODEM: Multi-Modal Digital Elevation Model Registration
 ![Registered Mesh](./docs/img/reg_mesh.png)
 
 
 ## Overview
-`MODEM` is a testbed application for registering a 3D model of an area of interest (AOI) to a larger 3D Foundation data source. Point cloud, mesh, and raster digital surface model (DSM) data types are supported. Format support is limited to the following:
+`CODEM` is a testbed application for registering a 3D model of an area of interest (AOI) to a larger 3D Foundation data source. Point cloud, mesh, and raster digital surface model (DSM) data types are supported. Format support is limited to the following:
 * Point Cloud: LAS, LAZ, BPF
 * Mesh: PLY, OBJ
 * DSM: GeoTIFF
 
-`MODEM` follows the following basic steps to perform co-registration:
+`CODEM` follows the following basic steps to perform co-registration:
 1. Generation of normalized DSMs from the AOI and Foundation data sources.
 2. Coarse registration via matching of features extracted from the DSMs.
 3. Fine registration via an iterative closest point (ICP) algorithm.
 4. Application of the solved registration transformation to the AOI data in its original type and format.
 
 
-## Installing MODEM
+## Installing CODEM
 1. Clone the repo:
 
     ```
-    git clone https://bitbucket.di2e.net/scm/crrelneggs/modem.git
-    ```
-    or
-    ```
-    git clone https://bitbucket.di2e.net/scm/crrelneggs/modem.git
+    git clone https://github.com/NCALM-UH/CODEM
     ```
 
-2. Create and activate a Conda environment containing the required dependencies. From inside the `modem` directory:
+
+2. Create and activate a Conda environment containing the required dependencies. From inside the `CODEM` directory:
 
     ```
     conda env create --file environment.yml
     ```
     ```
-    conda activate modem
+    conda activate codem
     ```
 
-3. Install `MODEM`. From inside the `modem` directory:
+3. Install `CODEM`. From inside the `codem` directory:
     ```
     pip install .
     ```
 
 
-## Running MODEM
-The `MODEM` application has two required positional arguments and numerous options. The required positional arguments are the file path to the Foundation data file and the file path to the AOI data file. Executing modem on the command line has the following form:
+## Running CODEM
+The `CODEM` application has two required positional arguments and numerous options. The required positional arguments are the file path to the Foundation data file and the file path to the AOI data file. Executing codem on the command line has the following form:
 ```
-modem <foundation_file_path> <aoi_file_path> [-opt option_value]
+codem <foundation_file_path> <aoi_file_path> [-opt option_value]
 ```
 
-For example, running `MODEM` on some of the sample data files in the [demo](demo) directory looks like:
+For example, running `CODEM` on some of the sample data files in the [demo](demo) directory looks like:
 ```
-modem demo/Foundation-PointCloud.laz demo/AOI-Mesh.ply
+codem demo/Foundation-PointCloud.laz demo/AOI-Mesh.ply
 ```
 
 Optional arguments can be placed before or after the positional arguments. For example, we can set the minimum registration pipeline resolution to a new value (default value = 1.0):
 ```
-modem demo/Foundation-PointCloud.laz demo/AOI-Mesh.ply --min_resolution 2.0
+codem demo/Foundation-PointCloud.laz demo/AOI-Mesh.ply --min_resolution 2.0
 ```
 
 A summary of all options and their default values is given in the [docs/configuration.md](docs/configuration.md) document. The default option values should be sufficient for most landscapes.
@@ -70,13 +67,15 @@ All output is saved to a new directory that is created at the location of the AO
 
 
 ## Additional Information
-Information on available configuration options, a more in-depth review of how `MODEM` works, and a simple example utilizing data files contained in the `demo` directory of this repository are found in the `docs` directory:
+Information on available configuration options, a more in-depth review of how `CODEM` works, and a simple example utilizing data files contained in the `demo` directory of this repository are found in the `docs` directory:
 * [docs/configuration.md](docs/configuration.md)
 * [docs/details.md](docs/details.md)
 * [docs/example.md](docs/example.md)
 
 
 ## Contact
+
+* Ognyan Moore - Hobu Inc. - [Email](ogi@hobu.co)
 * Preston Hartzell - University of Houston - [Email](pjhartzell@uh.edu)
 * Jesse Shanahan - formerly of Booz Allen Hamilton (listed for software development credit attribution) - [LinkedIn](https://www.linkedin.com/in/jesseshanahan/)
