+++ src/codem/__init__.py
@@ -1,0 +1,8 @@
+import codem.lib.log as log
+import codem.lib.resources as resources
+import codem.preprocessing.preprocess as preprocess
+import codem.registration.DsmRegistration as DsmRegistration
+import codem.registration.IcpRegistration as IcpRegistration
+import codem.registration.ApplyRegistration as ApplyRegistration
+import codem.main
+from codem.main import codemRunConfig, preprocess, coarse_registration, fine_registration, apply_registration
