+++ src/codem/lib/resources.py
@@ -1,0 +1,11 @@
+"""
+resources.py
+Project: CRREL-NEGGS University of Houston Collaboration
+Date: February 2021
+
+Supported filetypes
+"""
+
+dsm_filetypes = [".tif"]
+pcloud_filetypes = [".las", ".laz", ".bpf"]
+mesh_filetypes = [".ply", ".obj"]
