--- src/modem/__init__.py
@@ -1,8 +1,0 @@
-import modem.lib.log as log
-import modem.lib.resources as resources
-import modem.preprocessing.preprocess as preprocess
-import modem.registration.DsmRegistration as DsmRegistration
-import modem.registration.IcpRegistration as IcpRegistration
-import modem.registration.ApplyRegistration as ApplyRegistration
-import modem.main
-from modem.main import ModemRunConfig, preprocess, coarse_registration, fine_registration, apply_registration
